.
The new Pipelines with the 'optimized' models are crap. The defaults have been better. Or to be more precisely: The margin of the precision and accuracy gain in the optimized models is some where in the 2nd digit after the decimal point, where on the other hand some models even suffer from my 'improvement'.